refactor: tobe 패키지 삭제
