+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/ClassInstantiationException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 public class ClassInstantiationException extends IllegalArgumentException {
     public ClassInstantiationException(String message, Throwable cause) {
