+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/UnSupportedHandlerException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 public class UnSupportedHandlerException extends IllegalArgumentException {
     public UnSupportedHandlerException(String message) {
