+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/support/PathPatternUtilTest.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe.support;
+package com.interface21.webmvc.servlet.mvc.support;
 
 import org.junit.jupiter.api.Test;
